requirements.txt
Django
psycopg2-binary
pillow
djangorestframework
python-decouple
djangorestframework-simplejwt
pillow
django-vote